Analysis.CFG

Description

This module contains the data definitions and some utility functions of the Control Flow Graph (CFG).

Documentation

data CFGNodeValue

Data type containing the value of a node in the control flow graph.

Constructors

StatNode CompoundStmt'
A statement.

ForInitNode ForInit'
A for init.

ForUpdateNode Exps'
A for update.

CatchNode Catch'
A catch entry.

FinallyNode CompoundStmts'
A finally entry.

CallNode Scope Node Name'
An invocation of a method.

MethodEntryNode Scope
An entry of a method.

MethodExitNode Scope
An exit of a method.

type CFGNode = LNode CFGNodeValue

Type declaration of the node in the control flow graph.

type CFGNodes = [CFGNode]

Type declaration of the nodes in the control flow graph.

data CFGEdgeValue

Data type containing the value of an edge in the control flow graph.

Constructors

InterEdge Scope

IntraEdge

BlockEntryEdge BlockEntryType

BlockExitEdge BlockEntryType

BlockExitsEdge [BlockEntryType]

BlockExitEntryEdge BlockEntryType BlockEntryType

data BlockEntryType

Data type containing the type of entry or exit block.

Constructors

TryEntryType

CatchEntryType (Maybe FormalParam')

ConditionalEntryType (Maybe Exp')

FinallyEntryType

BlockEntryType (Maybe String)

type CFGEdge = LEdge CFGEdgeValue

Type declaration of the edge in the control flow graph.

type CFGEdges = [CFGEdge]

Type declaration of the edges in the control flow graph.

type CFGAdj = Adj CFGEdgeValue

Type declaration of the adjacencies in the control flow graph.

type CFGContext = (CFGAdj, Node, CFGNodeValue, CFGAdj)

Type declaration of the context in the control flow graph.

newtype CFG

Type declaration of the control flow graph.

Constructors

CFG

Fields

cfg :: Gr CFGNodeValue CFGEdgeValue

constructCFG :: CFGNodes -> CFGEdges -> CFG

Constructs the control flow graph from the given nodes and edges.

entryOfMethod :: Scope -> CFG -> Maybe CFGNode

Retruns the method entry node of the scope, if it exists.

isIntraEdge :: CFGEdgeValue -> Bool

Return true if the edge is an intraprocedural edge.

isInterEdge :: CFGEdgeValue -> Bool

Return true if the edge is an interprocedural edge.

Orphan instances

Eq (LNode a)

Methods

(==) :: LNode a -> LNode a -> Bool

(/=) :: LNode a -> LNode a -> Bool

Ord (LNode a)

Methods

compare :: LNode a -> LNode a -> Ordering

(<) :: LNode a -> LNode a -> Bool

(<=) :: LNode a -> LNode a -> Bool

(>) :: LNode a -> LNode a -> Bool

(>=) :: LNode a -> LNode a -> Bool

max :: LNode a -> LNode a -> LNode a

min :: LNode a -> LNode a -> LNode a