Auxiliary.Arguments

Description

Documentation

module Parsing.Syntax

module Linearization.Path

data Arguments

The arguments for the verification tool.

Constructors

Arguments

Fields

program :: FilePath

The path to the program to verify.

function :: Maybe String

The function to verify, default to main if it is Nothing.

verbosity :: Verbosity

The amount of information that is printed to the screen.

numberOfThreads :: Int

The number of threads to use when compiling and verifying.

removeOutputFiles :: Bool

The mark bool which states that the compiled files will be removed, if set to true.

maximumDepth :: Int

The maximum depth to generate program paths for.

pathMapFilter :: CompilationUnit' -> Maybe CompilationUnit'

The program path map filter function. When this function returns Nothing, for a specific path, this path will not be verified.

jbmcEnableAssertions :: Bool

Enable assertions in JBMC.

jbmcDepth :: Maybe Int

Sets the maximum depth in JBMC, if Just.

jbmcUnwind :: Maybe Int

Sets the maximum loop unwinding in JBMC, if Just.