Auxiliary.Phase

Description

Documentation

module Control.Monad.Trans.Except

module Control.Monad.IO.Class

module Auxiliary.Arguments

type PhaseResult a = ExceptT PhaseError IO a

Type declaration of a phase result, allowing for exceptions in the IO monad.

data PhaseError

Data type containing the different errors.

Constructors

ParsingError String

SemanticalError SemanticalError

SyntacticalError String

ResultError String

ExternalError String (Maybe String)

data SemanticalError

Data type containing the different semantical errors.

Constructors

UndefinedMethodReference Name'

UndefinedClassReference String

throwParsingError :: String -> PhaseResult a

Throws a parsing error.

throwSemanticalError :: SemanticalError -> PhaseResult a

Throws a semantical error.

throwSyntacticalError :: String -> PhaseResult a

Throws a syntactical error.

throwResultError :: String -> PhaseResult a

Throws a result error.

throwExternalError :: String -> Maybe String -> PhaseResult a

Throws an external error.

workingDir :: FilePath

The directory to compile and verify the program paths in.

type Phase a b = Arguments -> a -> PhaseResult b

Type declaration of a phase.

type Subphase a b = Phase a b

Type declaration of a subphase.