Auxiliary.Pretty

Description

This module contains the utilities to generate and print information to the screen.

The progress bar utility is exposed in this module as well, which can be used to render and update a progress bar.

Documentation

data Verbosity

Constructors

Compact

Informative

class Pretty a where

Minimal complete definition

pretty

Methods

pretty :: a -> Doc

toString :: a -> String

spaces :: Int -> Doc

newline :: Doc

newlines :: Doc -> Doc

tab :: Doc -> Doc

dot :: Doc

commas :: Pretty a => [a] -> Doc

dots :: Pretty a => [a] -> Doc

printHeader :: String -> IO ()

printPretty :: Pretty a => a -> IO ()

printTitled :: String -> String -> IO ()

printText :: String -> IO ()

type ProgressBar = ProgressBar ()

progressBar :: Int -> IO ProgressBar

incProgress :: ProgressBar -> IO ()