Linearization.Path

Description

Documentation

module Analysis.CFG

type ProgramPaths = [ProgramPath]

Type declaration of multiple program paths.

type ProgramPath = [PathStmt]

Type declaration of one program path.

type PathStmt = (PathType, PathMetaInfo)

Type declaration of a single statement, a block entry or a block exit, in a program path.

type PathMetaInfo = (Scope, String)

Type declaration of extra information of a PathStmt.

data PathType

Data type representing the a statement, a block entry, or a block exit.

Constructors

PathStmt Stmt'

PathEntry BlockEntryType

PathExit BlockEntryType