Parsing.Phase

Description

This module contains the parsing and syntax transformation phases. It parses the code of a given String and transforms it to the equivalent AST defined in Parsing.Syntax.

Documentation

parsingPhase :: Phase String CompilationUnit'

Parses the given string and transforms it to the locally declared AST.

syntaxTransformationSubphase :: Subphase CompilationUnit CompilationUnit'

Transforms the Language.Java.Syntax AST into the locally declared AST.