Parsing.Pretty

Description

This module contains the Pretty instances of the data types declared in Parsing.Syntax.

The Pretty instances generate compilable strings.

Documentation

preAndInlineModifiers :: (Modifiers', Modifiers') -> Modifiers' -> (Modifiers', Modifiers')

Splits the given modifiers in two sets, the first set are defined before the element and the second set are defined on the same line.

pForInit :: MaybeForInit' -> Doc

Pretty print a for init.

pForUpdate :: MaybeExps' -> Doc

Pretty print a for update.

pIdent :: Maybe String -> Doc

Pretty print an identifier.

Orphan instances

Pretty VarInit'

Methods

pretty :: VarInit' -> Doc

toString :: VarInit' -> String

Pretty VarDeclId'

Methods

pretty :: VarDeclId' -> Doc

toString :: VarDeclId' -> String

Pretty VarDecl'

Methods

pretty :: VarDecl' -> Doc

toString :: VarDecl' -> String

Pretty TypeDecls'

Methods

pretty :: TypeDecls' -> Doc

toString :: TypeDecls' -> String

Pretty TypeDecl'

Methods

pretty :: TypeDecl' -> Doc

toString :: TypeDecl' -> String

Pretty Type'

Methods

pretty :: Type' -> Doc

toString :: Type' -> String

Pretty SwitchBlocks'

Methods

pretty :: SwitchBlocks' -> Doc

toString :: SwitchBlocks' -> String

Pretty SwitchBlock'

Methods

pretty :: SwitchBlock' -> Doc

toString :: SwitchBlock' -> String

Pretty Stmt'

Methods

pretty :: Stmt' -> Doc

toString :: Stmt' -> String

Pretty Scope

Methods

pretty :: Scope -> Doc

toString :: Scope -> String

Pretty RefType'

Methods

pretty :: RefType' -> Doc

toString :: RefType' -> String

Pretty PrimType'

Methods

pretty :: PrimType' -> Doc

toString :: PrimType' -> String

Pretty Op'

Methods

pretty :: Op' -> Doc

toString :: Op' -> String

Pretty Modifier'

Methods

pretty :: Modifier' -> Doc

toString :: Modifier' -> String

Pretty MethodInvocation'

Methods

pretty :: MethodInvocation' -> Doc

toString :: MethodInvocation' -> String

Pretty MemberDecl'

Methods

pretty :: MemberDecl' -> Doc

toString :: MemberDecl' -> String

Pretty MaybeVarInits'

Methods

pretty :: MaybeVarInits' -> Doc

toString :: MaybeVarInits' -> String

Pretty Literal'

Methods

pretty :: Literal' -> Doc

toString :: Literal' -> String

Pretty Lhs'

Methods

pretty :: Lhs' -> Doc

toString :: Lhs' -> String

Pretty ImportDecls'

Methods

pretty :: ImportDecls' -> Doc

toString :: ImportDecls' -> String

Pretty ImportDecl'

Methods

pretty :: ImportDecl' -> Doc

toString :: ImportDecl' -> String

Pretty FormalParam'

Methods

pretty :: FormalParam' -> Doc

toString :: FormalParam' -> String

Pretty ForInit'

Methods

pretty :: ForInit' -> Doc

toString :: ForInit' -> String

Pretty FieldAccess'

Methods

pretty :: FieldAccess' -> Doc

toString :: FieldAccess' -> String

Pretty Exp'

Methods

pretty :: Exp' -> Doc

toString :: Exp' -> String

Pretty ElementValue'

Methods

pretty :: ElementValue' -> Doc

toString :: ElementValue' -> String

Pretty Decls'

Methods

pretty :: Decls' -> Doc

toString :: Decls' -> String

Pretty Decl'

Methods

pretty :: Decl' -> Doc

toString :: Decl' -> String

Pretty CompoundStmts'

Methods

pretty :: CompoundStmts' -> Doc

toString :: CompoundStmts' -> String

Pretty CompoundStmt'

Methods

pretty :: CompoundStmt' -> Doc

toString :: CompoundStmt' -> String

Pretty CompilationUnit'

Methods

pretty :: CompilationUnit' -> Doc

toString :: CompilationUnit' -> String

Pretty ClassType'

Methods

pretty :: ClassType' -> Doc

toString :: ClassType' -> String

Pretty ClassDecl'

Methods

pretty :: ClassDecl' -> Doc

toString :: ClassDecl' -> String

Pretty Catches'

Methods

pretty :: Catches' -> Doc

toString :: Catches' -> String

Pretty Catch'

Methods

pretty :: Catch' -> Doc

toString :: Catch' -> String

Pretty AssignOp'

Methods

pretty :: AssignOp' -> Doc

toString :: AssignOp' -> String

Pretty ArrayIndex'

Methods

pretty :: ArrayIndex' -> Doc

toString :: ArrayIndex' -> String

Pretty AnnotationKeyValues'

Methods

pretty :: AnnotationKeyValues' -> Doc

toString :: AnnotationKeyValues' -> String

Pretty AnnotationKeyValue'

Methods

pretty :: AnnotationKeyValue' -> Doc

toString :: AnnotationKeyValue' -> String

Pretty Annotation'

Methods

pretty :: Annotation' -> Doc

toString :: Annotation' -> String

Pretty [VarDecl']

Methods

pretty :: [VarDecl'] -> Doc

toString :: [VarDecl'] -> String

Pretty [Modifier']

Methods

pretty :: [Modifier'] -> Doc

toString :: [Modifier'] -> String

Pretty [FormalParam']

Methods

pretty :: [FormalParam'] -> Doc

toString :: [FormalParam'] -> String

Pretty (Maybe Type')

Methods

pretty :: Maybe Type' -> Doc

toString :: Maybe Type' -> String